# Performance improvements for `build_histogram` and `adjust_histogram`