--- a/test/exposure.jl
+++ b/test/exposure.jl
@@ -96,7 +96,7 @@ eye(m,n) = Matrix{Float64}(I,m,n)
                 @test axes(counts) == (0:length(edges),)
                 # Due to roundoff errors the bins are not the same as in the
                 # integer case above.
-                if T == Gray{Float32}
+                if T == Gray{Float32} || T == Gray{N0f8}
                     @test collect(counts) == [120, 5, 5, 5, 121]
                 else
                     @test collect(counts) == [120, 6, 5, 4, 121]
@@ -123,7 +123,7 @@ eye(m,n) = Matrix{Float64}(I,m,n)
                 img = 200/255:1/255:240/255
                 edges, counts  = build_histogram(T.(img),4,200/255,240/255)
                 @test length(edges) == length(counts) - 1
-                if T == Gray{Float32}
+                if T == Gray{Float32} || T == Gray{N0f8}
                     @test collect(counts) == [0, 10, 10, 10, 11]
                 else
                     @test collect(counts) == [0, 11, 9, 10, 11]
@@ -133,14 +133,8 @@ eye(m,n) = Matrix{Float64}(I,m,n)
                 edges, counts  = build_histogram(T.(img),4)
                 @test axes(counts) == (0:length(edges),)
                 @test length(edges) == length(counts) - 1
-                if T == Gray{N0f16} || T == Gray{N0f8}
-                    @test collect(counts) == [0, 11, 10, 10, 10]
-                elseif T == Gray{Float32}
-                    @test collect(counts) == [0, 10, 10, 10, 11]
-                else
-                    @test collect(counts) == [0, 11, 9, 10, 11]
-                end
-
+                @test all([0, 9, 9, 9, 9] .<= collect(counts) .<= [0, 11, 11, 11, 11])
+                @test sum(counts) == length(img)
             end
         end