@johnnychen94 2019-12-31
I feel our expectation on pushing Images.jl to v1.0 should be on API stabilization and deprecations, code rearrangement, documentation, and versioning. Maintainable work on performance optimization and multi-threads require more effort than we can pay in a short period (say, two months) and I view them as v1.x enhancements.