@timholy 2019-12-31
I agree that across-the-board performance enhancement is something to work on later. But while we have a couple of folks writing benchmarks, I want them to know that we can fix some of the problems they are observing. A subset of methods in JuliaImages have been very carefully optimized, but it doesn't attract contributors if we're too slow on common tasks.

And yes, I'm keenly interested in interop particularly with OpenCV. It's quite likely that the easiest way to systematically benchmark against OpenCV will be to have good wrappers.