@timholy 2019-12-31
Will ImageContrastAdjustment move to JuliaImages?