@zygmuntszpak 2019-12-31
What's the process involved in moving the repo to JuliaImages? I believe it makes sense to move it, but I'd like to understand what moving it entails and what the implied social contract is once the repo is in JuliaImages. For example, at the moment the Documenter script that builds the documentation is linked with my Github account using my credentials. I wouldn't know how to configure the equivalent in JuliaImages.