@johnnychen94 2019-12-31
As far as I understand, it's

1. [duplicate the repository](https://help.github.com/en/github/creating-cloning-and-archiving-repositories/duplicating-a-repository)  to JuliaImages by maintainers/owners
2. set up CI and Documenter under JuliaImages by maintainers/owners
3. submit a PR in the General registry to modify the package URL, which might need agreements on both sides.
4. post-operation: archive the original repository for reference usage, and optionally deactivate related CI by original owner.

I think step 2, 3, and 4 can be done parallelly. Please correct me if there's anything I missed to consider,