@kmsquire 2019-12-31
You should just be able to transfer the repo without duplicating it. Look
under the repo settings.

Steps 2 and 3 are still relevant.