@timholy 2019-12-31
> I'd like to understand what moving it entails and what the implied social contract is once the repo is in JuliaImages.

In terms of social contract it's probably no different than for anything else that Images.jl depends on. Once Images.jl moved to JuliaImages, I don't think it makes sense to have Images.jl depend on anything "image-processing specific" that isn't hosted at JuliaImages. That's just insurance that important dependencies won't get abandoned by someone who doesn't merge PRs and doesn't add collaborators with sufficient permissions (not that I expect to have to worry about that in your case, but it's a good policy).