@johnnychen94 2019-12-31
> You should just be able to transfer the repo without duplicating it

Yes, I thought the previous URL will become valid, but since it's actually get redirected to the new URL automatically, this isn't an issue. Transfering the whole repository is sufficient for step 1.

FYI, from https://help.github.com/en/github/administering-a-repository/transferring-a-repository#whats-transferred-with-a-repository

> All links to the previous repository location are automatically redirected to the new location.