@zygmuntszpak 2020-01-04
I've transferred the `ImageContrastAdjustment` repository to the JuliaImages organisation, and have opened a pull-request to the General registry which updates the repo URL (https://github.com/JuliaRegistries/General/pull/7492). @timholy I would be grateful if you could incorporate the changes you made in this pull-request in the ImageContrastAdjustment repo instead.