@timholy 2020-01-14
Thanks @zygmuntszpak! Would you like to do the honors of rebasing Images.jl on ImageContrastAdjustment? I also think that would be a good time to deprecate `imhist` and similar.