@zygmuntszpak 2020-01-14
I had actually started the process of updating the `Images` code on a local branch and paused because I noticed that `ImageContrastAdjustment` had a replacement for everything except the `clahe` function. Hence the new implementation of clahe https://github.com/JuliaImages/ImageContrastAdjustment.jl/pull/16

I am not entirely familiar with the deprecation strategy. Do I just reexport the functions in `ImageContrastAdjustment` and put a deprecation message in the existing `imhist` etc. functions?