@timholy 2020-01-14
Yep, that's pretty much it. You can use something along the lines of 
```julia
Base.depwarn("`imhist` will be removed in a future release, please use `build_histogram` instead.", :imhist)
```